Move event recorder
